Split the tests to make finding them easier
